Send sine wave data and plot it
The 'short' part of the sent data is a modified sine wave (modified to
fit into ints between 0 and 200). The rest of the data remains static
and unread by the front end.

When the data is received in the browser it is plotted onto a rolling
graph and displayed on screen.