Use caching argot and increase sendrate 4 > 100
This change pulls in the latest argot-browser that includes the caching
changes made in argot 0.1.9. These changes mean that the dictionaries do
not have to be repeatedly download and the libraries repeatedly
built.

The changes allow us to increase the data send rate from 4/second to
100/second, quite an increase. This also makes the graph a lot smoother.

One remaining problem is that still many load dictionary requests can be
made (for the same dictionary) before the cache is populated with that
dictionary's library representation. This means that the first requests
all race to get the dictionary (with a download each) until one of them
completes and populates the cache.

Perhaps some sort of loadDictionary queuing mechanism is required here.