Add all bower components
Heroku borks on the npm postinstall command that installs the bower
components, I'm not sure why but it seems to be a fairly common problem
among other users on the internet.

This change removes the bower_components directory from the ignore file
and adds them to the repo, it also edits the package.json file to remove
the troublesome postinstall command.