Update to Argot Browser 0.1.3
This change brings with it the ability to capture the processed library
after a `readMessage` and reuse it for subsequently read messages.

This change also seperates the setup and normal running phase where the
setup is downloading the dictionary and building the library from it,
and the normal running phase is receiving data and plotting it on the
graph.

This means that we are not registered to be notified of any data packets
until we have a built and cached library, thus removing the spikes we
were previously getting at the start of the process before. Win win.